refactor: extract permissions/branch-protection to github/permissions.py (#140)
