+++ src/mcp_server_git/github/api.py
@@ -2,9 +2,11 @@
 
 import logging
 
+from .ci_logs import *  # noqa: F401,F403
 from .client import github_client_context  # noqa: F401
 from .issues import *  # noqa: F401,F403
 from .patches import PatchMemoryManager  # noqa: F401  # re-export
+from .permissions import *  # noqa: F401,F403
 from .pr import *  # noqa: F401,F403
 from .pr_actions import *  # noqa: F401,F403
 from .releases import *  # noqa: F401,F403
